Lalonde Experimental Data

treat: Treatment assignment status (1 for treated, 0 for control)

age: Age of the individual

education: Years of education completed

black: Indicator for whether the individual is black

hispanic: Indicator for whether the individual is Hispanic

married: Indicator for whether the individual is married

nodegree: Indicator for whether the individual has no degree

re75: Earnings in 1975

re78: Earnings in 1978 (outcome variable)

Usage

data('nsw')

Value

A data frame with 722 observations and 9 variables

Details

Data from the National Supported Work Demonstration, a job training program studied by Lalonde (1986). The dataset includes both treated and control groups, with various demographic and earnings characteristics. This extract was created by Dehejia and Wahba (1999) and is commonly used in causal inference research.

References

"Causal Effects in Non-Experimental Studies: Reevaluating the Evaluation of Training Programs," Journal of the American Statistical Association, Vol. 94, No. 448 (December 1999), pp. 1053-1062.

"Propensity Score Matching Methods for Non-Experimental Causal Studies," Review of Economics and Statistics, Vol. 84, (February 2002), pp. 151-161.

Examples

data('nsw')