Sensitivity Summary (for survey weights)

Returns a data.frame or Kable table with summary measures of sensitivity; helper function for main summary function, and allows users to directly input a survey object and a design object

Usage

summarize_sensitivity_survey(svy_srs, svy_wt, weights, varY, b_star = 0)

Arguments

svy_srs

Survey object, containing the unweighted survey

svy_wt

Survey object, containing the weighted survey

weights

A vector, containing the estimated survey weights

varY

variance of the outcome

b_star

Killer confounder threshold, default set to be zero

Value

Sensitivity data.frame or latex table with summary measures of sensitivity

Examples

data(ces)
data(wapo)
wapo_srs <- survey::svydesign(ids = ~ 1, data = wapo)
#> Warning: No weights or probabilities supplied, assuming equal probability
ces_awt <- survey::svydesign(ids = ~ 1,
                     weights = ~ vvweight_post,
                     data = ces)

#Set up raking formula:
formula_rake <- ~ age_buckets + educ + gender + race + educ * pid + bornagain

#Generate targets:
targets_rake <- create_targets(ces_awt, formula_rake)

#PERFORM RAKING:
model_rake <- survey::calibrate(
  design = wapo_srs,
  formula = formula_rake,
  population = targets_rake,
  calfun = "raking",
  force = TRUE
)

rake_results <- survey::svydesign( ~ 1, data = wapo, weights = stats::weights(model_rake))
#Estimate from raking results:
weights = stats::weights(rake_results) * nrow(model_rake)
#Normalize to mean 1:
weights = weights / mean(weights)

unweighted_estimate = survey::svymean(~ candidate, wapo_srs, na.rm = TRUE)
weighted_estimate = survey::svymean(~ candidate, model_rake, na.rm = TRUE)
summarize_sensitivity(estimand = 'Survey',
Y = wapo$candidate,
weights = weights,
svy_srs = unweighted_estimate, 
svy_wt = weighted_estimate,
b_star = 0.53)
#>   Unweighted Unweighted_SE  Estimate         SE   RV
#> 1  0.5092251    0.01754357 0.4989841 0.01313149 0.08